refactor(e2e): ページでない AppSidebar を pom/components へ移す
サイドバーはページではなくコンポーネントのため、pom 直下ではなく
pom/components 配下に配置する。

refs #734
